Use non-NULL sentinel for boost::throws()
The contract for `boost::throws()` requires a function that advertises compliance to check whether the provided error code object is equivalent to the object represented by `boost::throws()`. Because `boost::throws()` is inlined, and because it returns a reference type, clang optimized this away in `clock_clock_p_d` because it is a reference being tested against `NULL`. The C++ spec states that references must be initialized, so the compiler is at liberty to make this optimization. This worked with the deprecated logic because there was a linked symbol to test. Replacing this with a non-NULL sentinel value allows this to work.